# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: reHome Support Module API (kb-support)
  version: 1.1.0
  description: 'API модуля службы поддержки reHome. Приём, обработка и сопровождение обращений пользователей силами операторов. Первая линия — AI-чат kb-search, эскалации приходят сюда.


    Архитектурные решения зафиксированы в ADR-0005:

    - Данные User/Booking/Premises/Collaborator — через API основной платформы

    - Ticket и ServiceOrder — разные сущности, связь по ссылке

    - Email — свой IMAP-парсер в контуре РФ

    - Уведомления: email + push + AI-чат + SMS (Exolve)

    - access_level на уровне хранилища, scope из JWT (ADR-0003)


    Стиль совпадает с основным API базы знаний: cursor-пагинация, RFC 7807 для ошибок, hybrid auth (m2m Bearer + browser cookie). v1.1: добавлены претензионные типы (COMPENSATION/GUARANTEE/INSURANCE/ACCEPTANCE_ACT), case_state и связанные поля, схема TicketCaseDetails, действия /decision и /case-state, события webhook case_decided/payout_released/insurance_event.'
  contact:
    name: ООО «РЕХОМ»
servers:
- url: https://api.rehome.one
  description: Production
- url: https://staging-api.rehome.one
  description: Staging
...
security:
- BearerAuth: []
- CookieAuth: []
paths:
  /api/v1/support/tickets:
    get:
      tags:
      - Tickets
      summary: Список заявок
      description: 'Возвращает заявки с фильтрами и cursor-пагинацией. Видимость зависит от scope: заявитель видит только свои; оператор — по своей команде/scope (фильтрация на уровне хранилища, ADR-0003).'
      operationId: listTickets
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/TicketStatus'
      - name: type
        in: query
        schema:
          $ref: '#/components/schemas/TicketType'
      - name: priority
        in: query
        schema:
          $ref: '#/components/schemas/TicketPriority'
      - name: channel
        in: query
        schema:
          $ref: '#/components/schemas/TicketChannel'
      - name: assignee_id
        in: query
        schema:
          type: string
          format: uuid
      - name: requester_id
        in: query
        schema:
          type: string
          format: uuid
      - name: team
        in: query
        schema:
          $ref: '#/components/schemas/TicketTeam'
      - name: premises_id
        in: query
        schema:
          type: string
          format: uuid
      - name: sla_breached
        in: query
        schema:
          type: boolean
      - name: tag
        in: query
        schema:
          type: string
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - created_at
          - -created_at
          - resolution_due_at
          - -resolution_due_at
          - priority
          - -priority
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/TicketSummary'
                    pagination:
                      $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Tickets
      summary: Создать заявку
      operationId: createTicket
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/support/tickets/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Tickets
      summary: Карточка заявки
      description: Заявитель видит только свою. Недоступная заявка → 404 (ADR-0003).
      operationId: getTicket
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Tickets
      summary: Обновить заявку
      description: Статус, приоритет, тип, команда, теги. Недопустимый переход статуса → 409.
      operationId: updateTicket
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/support/tickets/{id}/assign:
    post:
      tags:
      - Tickets
      summary: Назначить заявку
      operationId: assignTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - assignee_id
              properties:
                assignee_id:
                  type: string
                  format: uuid
                team:
                  oneOf:
                  - $ref: '#/components/schemas/TicketTeam'
                  - type: 'null'
  /api/v1/support/tickets/{id}/escalate:
    post:
      tags:
      - Tickets
      summary: Эскалировать на 2-ю линию
      operationId: escalateTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                team:
                  oneOf:
                  - $ref: '#/components/schemas/TicketTeam'
                  - type: 'null'
                reason:
                  type: string
  /api/v1/support/tickets/{id}/resolve:
    post:
      tags:
      - Tickets
      summary: Отметить решённой
      operationId: resolveTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                resolution_note:
                  type: string
  /api/v1/support/tickets/{id}/close:
    post:
      tags:
      - Tickets
      summary: Закрыть заявку
      operationId: closeTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/support/tickets/{id}/reopen:
    post:
      tags:
      - Tickets
      summary: Переоткрыть заявку
      operationId: reopenTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
  /api/v1/support/tickets/{id}/rate:
    post:
      tags:
      - Tickets
      summary: Оценка заявителя
      description: Доступно заявителю после RESOLVED/CLOSED. Оценка 1-2 уведомляет супервайзера.
      operationId: rateTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - rating
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Ticket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: m2m-доступ (например, от kb-search) — Bearer JWT через Client Credentials
    CookieAuth:
      type: apiKey
      in: cookie
      name: rehome_session
      description: Браузерная сессия оператора/заявителя. scope вычисляется из сессии на бэкенде.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Идентификатор запроса для трассировки
...
  schemas:
...
    TicketStatus:
      type: string
      enum:
      - NEW
      - OPEN
      - PENDING
      - WAITING
      - ESCALATED
      - RESOLVED
      - CLOSED
      - REOPENED
      description: NEW — новое; OPEN — в работе; PENDING — ждёт заявителя; WAITING — ждёт третью сторону (коллаборант/юрист/банк); ESCALATED — на 2-й линии; RESOLVED — решено, ждёт подтверждения; CLOSED — закрыто; REOPENED — переоткрыто
    TicketPriority:
      type: string
      enum:
      - low
      - normal
      - high
      - critical
...